nixos/tests/test-driver: add some clarifying docstrings & comments
