+++ nixos/lib/test-driver/test-driver.py
@@ -12,12 +12,14 @@ if __name__ == "__main__":
         help="re-use a VM state coming from a previous run",
         action="store_true",
     )
-    (cli_args, vm_scripts) = arg_parser.parse_known_args()
+    # 'rest' effectively are a list of paths to nix scripts that start the
+    # machines. They are passed as cli args after
+    (cli_args, rest) = arg_parser.parse_known_args()
 
     logger = Logger()
     if not cli_args.keep_vm_state:
         logger("Machine state will be reset. To keep it, pass --keep-vm-state")
-    driver = Driver(logger, vm_scripts, cli_args.keep_vm_state)
+    driver = Driver(logger, rest, cli_args.keep_vm_state)
 
     tic = time.time()
     driver.run_tests()
